Classifying images with CNN
Classifying Covid-19 positive and negative patients from X-ray images
Recent findings indicate the presence of COVID-19 infections in the patients with irregular findings on chest X-rays. There are many reports on this topic that include machine learning strategies for the identification of COVID-19 using chest X-rays.
This project uses radiological imaging to determine whether the scanned patient has COVID-19 or not.